task runs but s key to skip block fails
the behavior of the s key is quite erratic.
Pressing it will yield expected behavior on half of the sessions, roughly.

What I noticed is that, whenever the s key is caught as a skip signal, it is not
displayed on MATLAB console. However, when the key doesn't trigger the skip behavior,
the letter s gets displayed on the MATLAB console at each key press...

what remains to be done before actual experiment with patients is to do full
experiment on myself.